refactor: extract JiraExportProvider+ADF (#638 slice B)
Second slice of #638 (after #874 Networking). Moves the ADF
(Atlassian Document Format) payload models + hardLimit helper
into a dedicated extension file.

Moved (5 items):
- hardLimit(_ blocks:maxCharacters:) — widen private → internal (called by makeDescription in base)
- JiraDocument: Encodable — widen file-scope private → internal
- JiraBlock: Encodable (with paragraph/bulletList static factories) — widen
- JiraInline: Encodable (with text/listItem static factories) — widen
- JiraADFNode: Decodable — widen (used by JiraSearchIssueFields in base for now, will move with Duplicates slice)

Zero external Sources/ or Tests/ references to any of the 4 ADF
wire types — internal is safe.

Line-count delta:
- JiraExportProvider.swift: 1139 → 1033 (-106 lines)
- JiraExportProvider+ADF.swift (new): 114 lines
- xcodeproj: +4

Test coverage: JiraExportProviderTests passes unchanged.

Closes #890. Continues #638 after PR #877 (slice A Networking).
